tweaks
git-svn-id: https://svn.r-project.org/R/trunk@52136 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/R/Vignettes.R
@@ -218,7 +218,7 @@ function(package, dir, lib.loc = NULL, quiet = TRUE, clean = TRUE)
         file.remove(f %w/o% c(".", "..", pdfs, origfiles))
     }
 
-    file.remove(".build.timestamp")
+    if(file.exists(".build.timestamp")) file.remove(".build.timestamp")
     ## Might have been in origfiles ...
 
     invisible(NULL)
